+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Add device overriding config support. "Non-standard DP" devices have possibility to be supported now.
 - Add Camera support (`sp`). Thanks @ErrorErrorError for the contribution
 - Add Air Conditioner support (`kt`).
+- Add Air Conditioner Controller support (`ktkzq`).
 
 
 ### Changed